Refactoring: Unified the check for known systems for subsumes and memeberOf. Extracted mapping functionality for memeberOf.
